Refactor DataContext and add comment reactions to TaskCommentSection
