+++ src/components/ui/button.tsx
@@ -6,7 +6,6 @@ import { buttonVariants } from './button-variants';
 export interface ButtonProps
   extends React.ButtonHTMLAttributes<HTMLButtonElement>,
     VariantProps<typeof buttonVariants> {
-  asChild?: boolean;
 }
 
 const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
