+++ src/contexts/DataContextHooks.ts
@@ -1,0 +1,10 @@
+import { useContext } from 'react';
+import { DataContext } from './DataContextInstance';
+
+export function useData() {
+  const context = useContext(DataContext);
+  if (context === undefined) {
+    throw new Error('useData must be used within a DataProvider');
+  }
+  return context;
+}
