+++ src/contexts/DataContextInstance.ts
@@ -1,0 +1,4 @@
+import { createContext } from 'react';
+import { DataContextType } from './DataContextTypes';
+
+export const DataContext = createContext<DataContextType | undefined>(undefined);
